+++ pnet_datalink/src/bindings/winpcap.rs
@@ -13,8 +13,6 @@
 extern crate libc;
 extern crate winapi as win;
 
-use pnet_sys::sockets;
-
 #[repr(C)]
 pub struct _ADAPTER;
 pub type ADAPTER = _ADAPTER;
@@ -104,7 +102,7 @@ pub type PIP_ADAPTER_INFO = *mut _IP_ADAPTER_INFO;
 const MAX_DHCPV6_DUID_LENGTH: usize = 130;
 const MAX_DNS_SUFFIX_STRING_LENGTH: usize = 256;
 
-pub type LPSOCKADDR = *mut sockets::SockAddr;
+pub type LPSOCKADDR = *mut pnet_sys::SockAddr;
 
 #[repr(C)]
 pub struct _SOCKET_ADDRESS {
