+++ pnet_datalink/src/unix_interfaces.rs
@@ -96,7 +96,7 @@ fn sockaddr_to_network_addr(sa: *const libc::sockaddr) -> (Option<MacAddr>, Opti
 
             (Some(mac), None)
         } else {
-            let addr = pnet_sys::sockets::sockaddr_to_addr(
+            let addr = pnet_sys::sockaddr_to_addr(
                 mem::transmute(sa),
                 mem::size_of::<libc::sockaddr_storage>()
             );
@@ -130,7 +130,7 @@ fn sockaddr_to_network_addr(sa: *const libc::sockaddr) -> (Option<MacAddr>, Opti
 
             (Some(mac), None)
         } else {
-            let addr = pnet_sys::sockets::sockaddr_to_addr(
+            let addr = pnet_sys::sockaddr_to_addr(
                 mem::transmute(sa),
                 mem::size_of::<libc::sockaddr_storage>()
             );
